HTTP Options Detection
Detect HTTP Options but it is done in Driver Files.

Only solution is to set a PCB in main.c to call a callback function everytime a TCP Packet is received. Therefore should change to another sort of traffic detection.

ICMP Echo and Ping to Google DNS Server is working if device is DHCP Client to a Router connected to internet. However packet forwarding is scrapped.